backend/requirements.txt
fastapi
uvicorn[gunicorn]
openai
redis
better_profanity
pydantic
python-dotenv
requests
streamlit
streamlit_extras
